requirements.txt
# DDAS Dependencies
# Install: pip install -r requirements.txt

# Web framework
Flask>=3.0.0
Werkzeug>=3.0.0

# Auth & Security
PyJWT>=2.8.0
cryptography>=41.0.0

# HTTP
requests>=2.31.0

# File monitoring
watchdog>=4.0.0

# AI
google-generativeai>=0.7.0

# Env management
python-dotenv>=1.0.0

# Cloud Storage Integrations (optional, install as needed)
# boto3>=1.26.0           # AWS S3
# google-cloud-storage>=2.10.0  # Google Cloud Storage
# azure-storage-blob>=12.14.0   # Azure Blob Storage
# paramiko>=3.0.0         # SFTP Support

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Compression
zstandard>=0.21.0

# Production server (optional, recommended over Flask dev server)
gunicorn>=22.0.0

# Dev / testing
pytest>=8.0.0
pytest-flask>=1.3.0
